separate implementations
Common.h contains a lot of different Arduino API declarations - this splits their implementations into several files for clarity

the files begin with 'Common' to indicate their relation to the Common.h interface and are postfixed with a descriptive word to indicate their particular purpose
