+++ core-implement/CommonTiming.cpp
@@ -1,0 +1,64 @@
+/*
+Copyright (c) 2020 SparkFun Electronics
+
+Permission is hereby granted, free of charge, to any person obtaining a copy
+of this software and associated documentation files (the "Software"), to deal
+in the Software without restriction, including without limitation the rights
+to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+copies of the Software, and to permit persons to whom the Software is
+furnished to do so, subject to the following conditions:
+
+The above copyright notice and this permission notice shall be included in all
+copies or substantial portions of the Software.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+SOFTWARE.
+*/
+
+#include "mbed.h"
+
+#define PinMode Arduino_PinMode
+#include "core-api/api/Common.h"
+#undef PinMode
+
+#include "bridge/pins.h"
+
+#define standInFunc() printf("Stand-In for '%s' [file: %s, line: %d]\n", __FUNCTION__, __FILE__, __LINE__)
+
+// #if DEVICE_LPTICKER
+// static mbed::LowPowerTimer t;
+// #else
+static mbed::Timer t;
+// #endif
+
+using namespace std::chrono_literals;
+using namespace std::chrono;
+
+void initTimer(void){
+    t.start();
+}
+
+unsigned long millis(void){
+    return duration_cast<milliseconds>(t.elapsed_time()).count();
+}
+
+unsigned long micros(void){
+    return t.elapsed_time().count();
+}
+
+void delay(unsigned long ms){
+#ifndef NO_RTOS
+  rtos::ThisThread::sleep_for(ms * 1ms);
+#else
+  wait_us(ms * 1000);
+#endif
+}
+
+void delayMicroseconds(unsigned int us){
+    wait_us(us);
+}
